requirements.txt
python-telegram-bot>=21.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
httpx>=0.27.0
